Day - 201 Q3

+++ A_Sum_of_Odd_Integers.cpp
@@ -14,10 +14,10 @@ int main ()
 
     while(t--)
     {
-        int n,k;
+        ll n,k;
         cin>>n>>k;
 
-        int tp = n - (k*k);
+        ll tp = n - (k*k);
         if(n >= k*k && tp % 2 == 0)
         cout<<"YES"<<'\n';
         else
